Self-Checkout Software/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Stripe == 11.3.0
Flask-Admin==1.6.1
WTForms==2.3.3
sendgrid==6.7.1
pytest==8.3.4
pytest-flask==1.3.0